LCP: bugfix. communicator id overflow so reduced to int16_t
Global dynamic, which I dont know the use, was coded in unsigned int
which made created int overflow since in the header it is int16_t.

+++ src/MPC_Lowcomm/src/communicator.c
@@ -915,7 +915,7 @@ static inline void __communicator_id_factory_init(void)
 	int process_count = mpc_common_get_process_count();
 
 	/* We have the whole range except two */
-	unsigned int global_dynamic = ( ( (unsigned int )-1) - COMM_ID_TO_SKIP);
+	uint16_t global_dynamic = ( ( (uint16_t)-1) - COMM_ID_TO_SKIP);
 
 	/* Ensure it can be divided evenly */
 	if(global_dynamic % process_count)
